VOICE-IO-DEVICE/requirements.txt
toml
python-aiml
playsound==1.2.2
pyAudio
pyttsx3
SpeechRecognition
